+++ shareaza/CtrlMediaFrame.cpp
@@ -1,7 +1,7 @@
 //
 // CtrlMediaFrame.cpp
 //
-// Copyright (c) Shareaza Development Team, 2002-2012.
+// Copyright (c) Shareaza Development Team, 2002-2014.
 // This file is part of SHAREAZA (shareaza.sourceforge.net)
 //
 // Shareaza is free software; you can redistribute it
@@ -2194,7 +2194,7 @@ void CMediaFrame::UpdateNowPlaying(BOOL bEmpty)
 
 		// Strip extension
 		LPCTSTR szFilename = m_sNowPlaying;
-		m_sNowPlaying = m_sNowPlaying.Left( PathFindExtension( szFilename ) - szFilename );
+		m_sNowPlaying = m_sNowPlaying.Left( (int)( PathFindExtension( szFilename ) - szFilename ) );
 	}
 
 	CRegistry::SetString( _T("MediaPlayer"), _T("NowPlaying"), m_sNowPlaying );
