+++ shareaza/DCNeighbour.cpp
@@ -430,7 +430,7 @@ BOOL CDCNeighbour::OnChat(CDCPacket* pPacket)
 
 	if ( LPCSTR szMessage = strchr( (LPCSTR)pPacket->m_pBuffer, '>' ) )
 	{
-		int nNickLen = szMessage - (LPCSTR)pPacket->m_pBuffer - 1;
+		int nNickLen = (int)( szMessage - (LPCSTR)pPacket->m_pBuffer - 1 );
 		CString sNick( UTF8Decode( (LPCSTR)&pPacket->m_pBuffer[ 1 ], nNickLen ) );
 
 		if ( nNickLen > 0 && m_sNick != sNick )
