+++ shareaza/UPnPFinder.cpp
@@ -1,7 +1,7 @@
 //
 // UPnPFinder.cpp
 //
-// Copyright (c) Shareaza Development Team, 2002-2011.
+// Copyright (c) Shareaza Development Team, 2002-2014.
 // This file is part of SHAREAZA (shareaza.sourceforge.net)
 //
 // Shareaza is free software; you can redistribute it
@@ -776,7 +776,7 @@ HRESULT CUPnPFinder::InvokeAction(ServicePointer pService,
 
 	for( INT_PTR nArg = 0 ; nArg < nArgs ; nArg++ )
 	{
-		nPos = nArg + 1;
+		nPos = (LONG)( nArg + 1 );
 		hr = SafeArrayPutElement( psaArgs, &nPos, ppVars[ nArg ] );
 		if ( FAILED( hr ) ) return hr;
 	}
